remove single_execution_tuner
